+++ src/content-scripts/init/components.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import store from '../store/index';
 import { Provider } from 'react-redux';
-import App from '../containers/App';
+import App from '../components/App';
 
 export function injectComponents() {
   const documentContainer = document.createElement('div');
